tests/wpt/web-platform-tests/preload/download_resources.html

<!DOCTYPE html>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that preloaded resources are downloaded');
</script>
<link rel=preload href="resources/dummy.js" as=script>
<link rel=preload href="resources/dummy.css" as=style>
<link rel=preload href="resources/square.png" as=image>
<link rel=preload href="/fonts/CanvasTest.ttf" as=font crossorigin>
<link rel=preload href="/media/white.mp4" as=media>
<link rel=preload href="/media/sound_5.oga" as=media>
<link rel=preload href="/media/foo.vtt" as=media>
<link rel=preload href="resources/dummy.xml?foo=bar" as=foobarxmlthing>
<link rel=preload href="resources/dummy.xml">
<body>
<script src="resources/dummy.js?pipe=trickle(d5)"></script>
<script>
    window.addEventListener("load", t.step_func(function() {
        var entries = performance.getEntriesByType("resource");
        assert_equals(performance.getEntriesByType("resource").length, 11);
        t.done();
    }));
</script>
</body>
